+++ src/components/forum/ForumCategory.tsx
@@ -1,6 +1,4 @@
-import Widget from "../common/widgets/Widget";
 import type { BoardSummary } from "../../types/forum";
-import BoardSummaryView from "./boards/BoardSummaryView";
 
 const ForumCategory: React.FC<{ title: string; subBoards: BoardSummary[] }> = ({
   title,
